+++ packages/specs/src/components/Menu.spec.ts
@@ -78,7 +78,7 @@ export const MenuSpec: ComponentSpec<MenuProps> = {
       background: "{color.neutral}" as TokenRef,
       backgroundHover: "{color.neutral-hover}" as TokenRef,
       backgroundPressed: "{color.neutral-pressed}" as TokenRef,
-      text: "{color.white}" as TokenRef,
+      text: "{color.base}" as TokenRef,
       border: "{color.neutral}" as TokenRef,
     },
     secondary: {